[DLMED] construct class
