+++ monai/bundle/utils.py
@@ -12,21 +12,21 @@
 import json
 import re
 from distutils.util import strtobool
-from typing import Any, Dict, Optional, Sequence, Tuple, Union
+from typing import Any, Dict, Tuple, Union
 
 from monai.bundle.config_parser import ConfigParser
-from monai.utils import ensure_tuple, optional_import
+from monai.utils import optional_import
 
 yaml, _ = optional_import("yaml")
 
 
 def load_config_file(filepath: str, **kwargs):
     """
-    Load config file with specified file path.
+    Load structured config file with the specified file path.
     Suppprt JSON and YAML formats.
 
     Args:
-        filepath: path of target file to load, supported postfixes: `.json`, `yml`, `yaml`.
+        filepath: path of target file to load, supported postfixes: `.json`, `.yml`, `.yaml`.
         kwargs: other arguments for `json.load` or `yaml.load`, depends on file format.
             for more details, please check:
             https://docs.python.org/3/library/json.html#json.load.
@@ -44,6 +44,7 @@ def load_config_file(filepath: str, **kwargs):
 def load_config_file_content(path: str, **kwargs):
     """
     Load part of the content from a config file with specified `id` in the path.
+    If no `id` provided, load the whole content of the file.
     Suppprt JSON and YAML formats file.
 
     Args:
@@ -66,33 +67,13 @@ def load_config_file_content(path: str, **kwargs):
     return parser[paths[1][1:] if paths[1] != "" else ""]
 
 
-def update_default_args(args: Optional[Union[str, Dict]] = None, **kwargs) -> Dict:
-    """
-    Update the `args` with the input `kwargs`.
-    For dict data, recursively update the content based on the keys.
-
-    Args:
-        args: source args to update.
-        kwargs: destination args to update.
-
-    """
-    args_: Dict = args if isinstance(args, dict) is None else {}  # type: ignore
-    if isinstance(args, str):
-        args_ = load_config_file_content(args)
-
-    # recursively update the default args with new args
-    for k, v in kwargs.items():
-        args_[k] = update_default_args(args_[k], **v) if isinstance(v, dict) and isinstance(args_.get(k), dict) else v
-    return args_
-
-
 def parse_id_value(pair: str) -> Tuple[str, Any]:
     """
     Parse the "id:value" pair string to `id` and `value`.
     Will try to convert the correct data type of `value` from string.
 
     Args:
-        pair (str): input "id:value" pair to parse.
+        pair: input "id:value" pair to parse.
 
     """
     items = pair.split(":")
@@ -121,49 +102,7 @@ def id_value_str_to_dict(pairs: str) -> Dict[str, Any]:
     """
     Utility to convert a string which represents a dict of `id:value` pairs to a python dict. For example:
     `"{postprocessing#<args>#postfix: output, network: <file>other.json#net_args}"`
-    Will try to convert the correct data type of `value` from string.
+    Will try to convert the data type of `value` from string to real type.
 
     """
-    return dict(map(parse_id_value, pairs[1:-1].split(",")))
-
-
-def parse_config_file(
-    config_file: Union[str, Sequence[str]], meta_file: Union[str, Sequence[str]], override: Optional[Dict] = None
-) -> ConfigParser:
-    """
-    Read the config file, metadata file and override with specified `id=value` pairs.
-    Put metadata in the config content with key "<meta>".
-    The `id` in `override` identifies target position to override with the `value`.
-    If `value` starts with "<file>", it will automatically read the `file`
-    and use the content as `value`.
-
-    Args:
-        config_file: filepath of the config file, the config content must be a dictionary,
-            if providing a list of files, wil merge the content of them.
-        meta_file: filepath of the metadata file, the config content must be a dictionary,
-            if providing a list of files, wil merge the content of them.
-        override: dict of `{id: value}` pairs to override or add the config content.
-
-    """
-    config: Dict = {"<meta>": {}}
-    for f in ensure_tuple(config_file):
-        content = load_config_file(f)
-        if not isinstance(content, dict):
-            raise ValueError("input config content must be a dictionary.")
-        config.update(content)
-
-    for f in ensure_tuple(meta_file):
-        content = load_config_file(f)
-        if not isinstance(content, dict):
-            raise ValueError("meta data content must be a dictionary.")
-        config["<meta>"].update(content)
-
-    parser = ConfigParser(config=config)
-
-    if override is not None:
-        for id, v in override.items():
-            if isinstance(v, str) and v.startswith("<file>"):
-                v = load_config_file_content(v[6:])
-            parser[id] = v
-
-    return parser
+    return dict(map(parse_id_value, pairs[1: -1].split(",")))
