+++ tests/test_bundle_run.py
@@ -135,8 +135,8 @@ class TestBundleRun(unittest.TestCase):
                 [
                     f"{sys.executable}",
                     "-m",
-                    "monai.bundle.scripts",
-                    "run",
+                    "monai.bundle.script",
+                    "Run",
                     "--meta_file",
                     f"{meta_file}",
                     "--config_file",
@@ -160,7 +160,7 @@ class TestBundleRun(unittest.TestCase):
                         "-m",
                         "fire",
                         "monai.bundle",
-                        "run",
+                        "Run",
                         "--meta_file",
                         f"{meta_file}",
                         "--config_file",
